+++ test/ResponseTest.php
@@ -1,0 +1,21 @@
+<?php
+
+use Stateless\Response;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @covers Response
+ */
+final class ResponseTest extends TestCase {
+    public function testHeader() {
+        //Response::header(404);
+    }
+    
+    public function testRedirect() {
+        //Response::redirect("/");
+    }
+
+    public function testSendToken() {
+        //Response::sendToken("");
+    }
+}